requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
qdrant-client==1.16.1
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
black==23.11.0
flake8==6.1.0
langchain==0.0.345
textblob==0.18.0.post0
pytest==7.4.3
httpx==0.25.2
PyJWT==2.8.0
email-validator==2.1.1
pydantic-settings==2.2.1
passlib==1.7.4
bcrypt
psycopg2-binary==2.9.9
python-jose==3.3.0